move resource-container-helper to utils
